requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
install==1.3.5
loguru==0.6.0
magic-filter==1.0.9
multidict==6.0.4
packaging==22.0
pip-review==1.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.7
toml==0.10.2
tomli==2.0.1
wgconfig==0.3.0
win32-setctime==1.1.0
yarl==1.8.2
